ExecutionPlan handler simplified
