+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/ImmutableTaskGraph.java
@@ -170,10 +170,6 @@ public class ImmutableTaskGraph {
         taskGraph.enableProfiler(profilerMode);
     }
 
-    void disableProfiler(ProfilerMode profilerMode) {
-        taskGraph.disableProfiler(profilerMode);
-    }
-
     void withConcurrentDevices() {
         taskGraph.withConcurrentDevices();
     }
