+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/ExecutionPlanType.java
@@ -3,13 +3,30 @@ package uk.ac.manchester.tornado.api.plantype;
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
 public sealed class ExecutionPlanType extends TornadoExecutionPlan //
-        permits OffConcurrentDevices, OffMemoryLimit, OffPrintKernel, OffProfiler, OffThreadInfo, WithWarmUp, WithBatch, WithClearProfiles, WithCompilerFlags, WithConcurrentDevices,
-        WithDefaultScheduler, WithDevicePlan, WithDynamicReconfiguration, WithFreeDeviceMemory, WithGridScheduler, WithMemoryLimit, WithPrintKernel, WithProfiler, WithResetDevice, WithThreadInfo {
+        permits OffConcurrentDevices, OffMemoryLimit, OffPrintKernel, //
+        OffProfiler, OffThreadInfo, WithWarmUp, WithBatch, WithClearProfiles,  //
+        WithCompilerFlags, WithConcurrentDevices, WithDefaultScheduler, //
+        WithDevicePlan, WithDynamicReconfiguration, WithFreeDeviceMemory, //
+        WithGridScheduler, WithMemoryLimit, WithPrintKernel, WithProfiler, //
+        WithResetDevice, WithThreadInfo {
 
-    TornadoExecutionPlan parent;
+    TornadoExecutionPlan parentLink;
 
     public ExecutionPlanType(TornadoExecutionPlan parent) {
-        this.parent = parent;
+
+        // Set link between the previous action (parent) and the new one
+        this.parentLink = parent;
+
+        // Copy the reference for the executor
+        this.tornadoExecutor = parent.getTornadoExecutor();
+
+        // Copy the reference for the execution frame
+        this.executionFrame = parent.getExecutionFrame();
+
+        // Copy the reference for the executor list
+        this.executorList = parent.getExecutionList();
+
+        // Set child reference to this instance
         this.child = this;
     }
