+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/OffConcurrentDevices.java
@@ -10,7 +10,7 @@ public final class OffConcurrentDevices extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withoutConcurrentDevices ";
+        return parentLink.toString() + "\n -> withoutConcurrentDevices ";
     }
 
 }
