+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/OffProfiler.java
@@ -10,7 +10,7 @@ public final class OffProfiler extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withoutProfiler ";
+        return parentLink.toString() + "\n -> withoutProfiler ";
     }
 
 }
