+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithBatch.java
@@ -10,6 +10,6 @@ public final class WithBatch extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return super.toString() + "\n -> withBatch ";
+        return parentLink.toString() + "\n -> withBatch ";
     }
 }
