+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithDefaultScheduler.java
@@ -10,6 +10,6 @@ public final class WithDefaultScheduler extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withDefaultGridScheduler ";
+        return parentLink.toString() + "\n -> withDefaultGridScheduler ";
     }
 }
