+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithDevicePlan.java
@@ -10,6 +10,6 @@ public final class WithDevicePlan extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withDevice ";
+        return parentLink.toString() + "\n -> withDevice ";
     }
 }
