+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithMemoryLimit.java
@@ -10,7 +10,7 @@ public final class WithMemoryLimit extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withMemoryLimit ";
+        return parentLink.toString() + "\n -> withMemoryLimit ";
     }
 
 }
