+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithPrintKernel.java
@@ -9,6 +9,6 @@ public final class WithPrintKernel extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withPrintKernel ";
+        return parentLink.toString() + "\n -> withPrintKernel ";
     }
 }
