+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithProfiler.java
@@ -10,6 +10,6 @@ public final class WithProfiler extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withProfiler ";
+        return parentLink.toString() + "\n -> withProfiler ";
     }
 }
