+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plantype/WithResetDevice.java
@@ -10,6 +10,6 @@ public final class WithResetDevice extends ExecutionPlanType {
 
     @Override
     public String toString() {
-        return parent.toString() + "\n -> withResetDevice ";
+        return parentLink.toString() + "\n -> withResetDevice ";
     }
 }
